Refactor render middleware into separate function (further de-duplication)
